requirements.txt
# Core Framework
Flask==3.0.0
Flask-Cors==4.0.0
python-dotenv==1.0.0
gunicorn==21.2.0
utils

# Angel One API
smartapi-python==1.3.5
pyotp==2.9.0
websocket-client==1.8.0

# AWS Services
boto3==1.34.0
awswrangler==3.5.0

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# PySpark (Compatible with Python 3.11.9)
pyspark==3.5.0
pyarrow==14.0.1

# Data Processing
pandas==2.1.4
numpy==1.26.2

# Machine Learning
scikit-learn==1.3.2
prophet==1.1.5
statsmodels==0.14.1

# Sentiment Analysis
textblob==0.17.1
nltk==3.8.1

# Web Scraping
beautifulsoup4==4.12.2
requests==2.31.0
lxml==5.0.0

# News API
newsapi-python==0.2.7

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
mplfinance==0.12.10b0
plotly==5.18.0

# PDF Generation
reportlab==4.0.7
Pillow==10.1.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Testing (Optional for now)
pytest==7.4.3
pytest-cov==4.1.0
